Move LRUCache to picard/utils/lrucache.py
Moved the LRUCache class to its own module in utils, in case it's useful
in other parts of the code.
